+++ src/pytraits/__init__.py
@@ -22,6 +22,6 @@ from pytraits.extendable import extendable
 from pytraits.setproperty import setproperty
 from pytraits.trait_composer import add_traits
 
-__version__ = "1.0.1"
+__version__ = "1.1.0"
 __all__ = ["Singleton", "combine_class", "extendable", "add_traits",
            "type_safe", "type_converted", "setproperty"]